refactor: split onboard web search selection
